docs: reorder js files
