+++ src/listeners.js
@@ -1,8 +1,12 @@
+import {getImage} from './photo';
+import {toggleMetric} from './display';
+
 const clickFahrenheit = async () => {
   try {
     const newcity = document.getElementById('city-title').textContent;
     const callFar = await toggleMetric('imperial', newcity);
-  } catch {
+  } catch (error) {
+    console.error(error);
     alert("Error: there's no informationa available yet.")
   }
 }
@@ -11,7 +15,8 @@ const clickCelsius = async () => {
   try {
     const newcityCel = document.getElementById('city-title').textContent;
     const callCel = await toggleMetric('metric', newcityCel);
-  } catch {
+  } catch (error) {
+    console.error(error);
     alert("Error: there's no informationa available yet.")
   }
 }
@@ -22,7 +27,10 @@ const clickSearch = async () => {
     const repo = await getImage(city);
     const call = await toggleMetric('metric', city);
     document.getElementById('inlineFormInputName2').value = '';
-  } catch {
+  } catch (error) {
+    console.error(error);
     alert("Sorry: we couldn't find your city");
   }
 }
+
+export {clickFahrenheit, clickCelsius, clickSearch};
